selenium_pytest/requirements.txt
selenium
pytest
allure-pytest
python-dotenv==1.0.0
